Make sure appmenu button appears on the Left
This may well be Linux/Fedora specific, but using oneLiner meant
that the appmenu button (the thing you get if you don't show the
menu bar) always displayed inbetween the urlbar and the tabs. And
to me that just looked stupid. So this forces it to the far left.

No idea what happens on other platforms with these changes.